asdf
git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@1641 29311d96-e01e-0410-9327-a35deaab8ce9

+++ branches/sage/mds/test/test_seek_read.c
@@ -11,7 +11,7 @@
 int main(int argc, char **argv)
 {
   char *fn = argv[1];
-  int numblocks = atoi(argv[2]) / 4096;
+  uint64_t numblocks = atoll(argv[2]) / 4096;
   int count = 400;
   
   cout << "fn " << fn << endl;
